rename domain folder
