+++ src/hero/application/hero.service.ts
@@ -1,11 +1,11 @@
 import { Injectable } from '@nestjs/common';
 import { CreateHeroDto } from '../dto/create-hero.dto';
 import { UpdateHeroDto } from '../dto/update-hero.dto';
-import { Hero } from '../domaine/hero';
+import { Hero } from '../domain/hero';
 import { HeroRepositoryNosql } from '../persistance/hero.repository.nosql';
-import { HeroWithoutId } from '../domaine/hero.without.id';
-import { HeroSpecialties } from '../domaine/hero.specialties';
-import { HeroRarities } from '../domaine/hero.rarities';
+import { HeroWithoutId } from '../domain/hero.without.id';
+import { HeroSpecialties } from '../domain/hero.specialties';
+import { HeroRarities } from '../domain/hero.rarities';
 
 @Injectable()
 export class HeroService {
