+++ src/hero/domain/hero.props.ts
