+++ src/hero/domain/hero.rarities.ts
