+++ src/hero/domain/hero.specialties.ts
