+++ src/hero/domain/hero.ts
