+++ src/hero/domain/hero.without.id.ts
